fix: iframe not resizing on load
Because iframe.js and index.js functions are all put on window,
iframe.js was actually calling postMessage function in index.js,
which clobbered the actual window.postMessage. The parent function
just then sent the message back to the child, so it was not receiving
any events.

Fixed by renaming parent function. All code should be in IIFEs to
prevent this whole class of errors, but this broke the code entirely and so would need to investigate
further for a proper fix.