switch to webcomponent
The singleton iframe pattern didn't work as is with SPA / static sites
as new elements were created with the same id, orphaning the one created
by the widget.

With this pattern, once the script is loaded the client can use a
cusdis-widget web component, that encapsulates setup and teardown and
can be used in multiple points throughout the app.